+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [1.12.3] - 2019-08-07
+### Added
+- `showUnder` and `duration` props to `scrollToTop` component
+### Changed
+- `ripple` transition timing
+
 ## [1.12.1] - 2019-08-05
 ### Fixed
 - `fluid` prop of the `button` component